face_detection_module/requirements.txt
absl-py==2.3.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
basicsr==1.4.2
certifi==2025.4.26
charset-normalizer==2.1.1
cmake==4.0.2
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
facexlib==0.2.5
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.27.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lit==18.1.8
llvmlite==0.43.0
lmdb==1.3.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==1.58.1
opencv-python==4.6.0.66
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0
realesrgan==0.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tb-nightly==2.20.0a20250531
tensorboard-data-server==0.7.2
tifffile==2025.5.26
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
triton==2.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==1.26.20
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.20.0
